new .hpp and .cpp files for input box class
