+++ CMakeLists.txt
@@ -13,7 +13,8 @@ find_package(SQLite3)
 set(SRC_LIST src/main.cpp src/sql.cpp src/game.cpp src/scene.cpp src/level_map_scene.cpp src/wardrobe_scene.cpp
  src/main_menu_scene.cpp src/level_scene.cpp src/player.cpp src/Level.cpp 
  src/TileMap.cpp src/hse_utils.cpp src/coin.cpp src/level_icon.cpp src/authentication_scene.cpp
- src/login_scene.cpp src/register_scene.cpp src/enemy.cpp src/lose_scene.cpp src/win_scene.cpp)
+ src/login_scene.cpp src/register_scene.cpp src/enemy.cpp src/lose_scene.cpp src/win_scene.cpp
+ src/input_box.cpp)
 
 include_directories(super_hse include)
